feat(dev): combined helpers, moved XLoaderFormats class;
- Combined helpers and conditions into one helper.
- Moved `XLoaderFormats` class to utils script to prevent circular imports.
