+++ ckanext/xloader/templates/package/resource_edit_base.html
@@ -2,10 +2,7 @@
 
 {% block inner_primary_nav %}
     {{ super() }}
-    {%
-      if (h.is_xloader_format(res.format) or res.datastore_active)
-      and h.is_xloader_type(res.url_type)
-    %}
+    {% if h.is_resource_supported_by_xloader(res) %}
       {{ h.build_nav_icon('xloader.resource_data', _('DataStore'), id=pkg.name, resource_id=res.id, icon='cloud-upload') }}
     {% endif %}
 {% endblock %}
