+++ ckanext/xloader/templates/package/snippets/resource_item.html
@@ -2,11 +2,7 @@
 
 {% block resource_item_explore_inner %}
   {{ super() }}
-  {%
-    if (h.is_xloader_format(res.format) or res.datastore_active)
-    and h.check_access('package_update', {'id':pkg.id })
-    and h.is_xloader_type(res.url_type)
-  %}
+  {% if h.is_resource_supported_by_xloader(res) %}
     <li>{% link_for _('DataStore'), named_route='xloader.resource_data', id=pkg.name, resource_id=res.id, class_='dropdown-item', icon='cloud-upload' %}</li>
   {% endif %}
 {% endblock %}
