+++ ckanext/xloader/templates/package/snippets/resources.html
@@ -2,11 +2,7 @@
 
 {% block resources_list_edit_dropdown_inner %}
   {{ super() }}
-  {%
-    if (h.is_xloader_format(resource.format) or resource.datastore_active)
-    and h.check_access('package_update', {'id':pkg.id })
-    and h.is_xloader_type(resource.url_type)
-  %}
+  {% if h.is_resource_supported_by_xloader(resource) %}
     <li>{% link_for _('DataStore'), named_route='xloader.resource_data', id=pkg.name, resource_id=resource.id, class_='dropdown-item', icon='cloud-upload' %}</li>
   {% endif %}
 {% endblock %}
